+++ smartcontract/cli/src/device/interface/update.rs
@@ -3,7 +3,7 @@ use crate::{
     doublezerocommand::CliCommand,
     poll_for_activation::poll_for_device_interface_activated,
     requirements::{CHECK_BALANCE, CHECK_ID_JSON},
-    validators::{validate_parse_bandwidth, validate_pubkey_or_code},
+    validators::validate_pubkey_or_code,
 };
 use clap::Args;
 use doublezero_program_common::validate_iface;
@@ -23,24 +23,25 @@ pub struct UpdateDeviceInterfaceCliCommand {
     /// Loopback type (if applicable)
     #[arg(long)]
     pub loopback_type: Option<types::LoopbackType>,
-    /// Interface CYOA
-    #[arg(long)]
-    pub interface_cyoa: Option<types::InterfaceCYOA>,
-    /// DIA Port (for DIA interfaces)
-    #[arg(long)]
-    pub interface_dia: Option<types::InterfaceDIA>,
-    /// Bandwidth in Mbps
-    #[arg(long, value_parser = validate_parse_bandwidth)]
-    pub bandwidth: Option<u64>,
-    /// Committed Information Rate in Mbps
-    #[arg(long, value_parser = validate_parse_bandwidth)]
-    pub cir: Option<u64>,
-    /// MTU
-    #[arg(long)]
-    pub mtu: Option<u16>,
-    /// Routing mode
-    #[arg(long)]
-    pub routing_mode: Option<types::RoutingMode>,
+    // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+    // /// Interface CYOA
+    // #[arg(long)]
+    // pub interface_cyoa: Option<types::InterfaceCYOA>,
+    // /// DIA Port (for DIA interfaces)
+    // #[arg(long)]
+    // pub interface_dia: Option<types::InterfaceDIA>,
+    // /// Bandwidth in Mbps
+    // #[arg(long, value_parser = validate_parse_bandwidth)]
+    // pub bandwidth: Option<u64>,
+    // /// Committed Information Rate in Mbps
+    // #[arg(long, value_parser = validate_parse_bandwidth)]
+    // pub cir: Option<u64>,
+    // /// MTU
+    // #[arg(long)]
+    // pub mtu: Option<u16>,
+    // /// Routing mode
+    // #[arg(long)]
+    // pub routing_mode: Option<types::RoutingMode>,
     /// VLAN ID (default: 0, i.e. not set)
     #[arg(long)]
     pub vlan_id: Option<u16>,
@@ -107,12 +108,13 @@ impl UpdateDeviceInterfaceCliCommand {
             pubkey: device_pk,
             name: self.name.clone(),
             loopback_type: self.loopback_type.map(|lt| lt.into()),
-            interface_cyoa: self.interface_cyoa.map(|ic| ic.into()),
-            interface_dia: self.interface_dia.map(|id| id.into()),
-            bandwidth: self.bandwidth,
-            cir: self.cir,
-            mtu: self.mtu,
-            routing_mode: self.routing_mode.map(|rm| rm.into()),
+            // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+            // interface_cyoa: self.interface_cyoa.map(|ic| ic.into()),
+            // interface_dia: self.interface_dia.map(|id| id.into()),
+            // bandwidth: self.bandwidth,
+            // cir: self.cir,
+            // mtu: self.mtu,
+            // routing_mode: self.routing_mode.map(|rm| rm.into()),
             vlan_id: self.vlan_id,
             user_tunnel_endpoint: self.user_tunnel_endpoint,
             status: self.status.as_ref().map(|s| s.parse().unwrap()),
@@ -136,7 +138,7 @@ mod tests {
     use crate::tests::utils::create_test_client;
     use doublezero_sdk::{AccountType, CurrentInterfaceVersion, Device, DeviceStatus, DeviceType};
     use doublezero_serviceability::state::interface::{
-        InterfaceCYOA, InterfaceStatus, InterfaceType, LoopbackType, RoutingMode,
+        InterfaceStatus, InterfaceType, LoopbackType,
     };
     use mockall::predicate;
     use solana_sdk::{pubkey::Pubkey, signature::Signature};
@@ -170,12 +172,13 @@ mod tests {
                     name: "Ethernet0".to_string(),
                     interface_type: InterfaceType::Physical,
                     loopback_type: LoopbackType::None,
-                    interface_cyoa: InterfaceCYOA::None,
-                    interface_dia: doublezero_serviceability::state::interface::InterfaceDIA::None,
-                    bandwidth: 1000,
-                    cir: 500,
-                    mtu: 1500,
-                    routing_mode: RoutingMode::Static,
+                    // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+                    // interface_cyoa: InterfaceCYOA::None,
+                    // interface_dia: doublezero_serviceability::state::interface::InterfaceDIA::None,
+                    // bandwidth: 1000,
+                    // cir: 500,
+                    // mtu: 1500,
+                    // routing_mode: RoutingMode::Static,
                     vlan_id: 0,
                     ip_net: "10.0.0.1/24".parse().unwrap(),
                     node_segment_idx: 0,
@@ -187,12 +190,13 @@ mod tests {
                     name: "Loopback0".to_string(),
                     interface_type: InterfaceType::Loopback,
                     loopback_type: LoopbackType::Vpnv4,
-                    interface_cyoa: InterfaceCYOA::None,
-                    interface_dia: doublezero_serviceability::state::interface::InterfaceDIA::None,
-                    bandwidth: 1000,
-                    cir: 500,
-                    mtu: 1500,
-                    routing_mode: RoutingMode::Static,
+                    // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+                    // interface_cyoa: InterfaceCYOA::None,
+                    // interface_dia: doublezero_serviceability::state::interface::InterfaceDIA::None,
+                    // bandwidth: 1000,
+                    // cir: 500,
+                    // mtu: 1500,
+                    // routing_mode: RoutingMode::Static,
                     vlan_id: 16,
                     ip_net: "10.0.1.1/24".parse().unwrap(),
                     node_segment_idx: 0,
@@ -221,12 +225,13 @@ mod tests {
                 pubkey: device1_pubkey,
                 name: "Loopback0".to_string(),
                 loopback_type: Some(LoopbackType::Ipv4),
-                interface_cyoa: None,
-                interface_dia: None,
-                bandwidth: None,
-                cir: None,
-                mtu: None,
-                routing_mode: None,
+                // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+                // interface_cyoa: None,
+                // interface_dia: None,
+                // bandwidth: None,
+                // cir: None,
+                // mtu: None,
+                // routing_mode: None,
                 vlan_id: Some(20),
                 user_tunnel_endpoint: None,
                 status: Some(InterfaceStatus::Activated),
@@ -242,12 +247,13 @@ mod tests {
             pubkey_or_code: device1_pubkey.to_string(),
             name: "Loopback0".to_string(),
             loopback_type: Some(types::LoopbackType::Ipv4),
-            interface_cyoa: None,
-            interface_dia: None,
-            bandwidth: None,
-            cir: None,
-            mtu: None,
-            routing_mode: None,
+            // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+            // interface_cyoa: None,
+            // interface_dia: None,
+            // bandwidth: None,
+            // cir: None,
+            // mtu: None,
+            // routing_mode: None,
             vlan_id: Some(20),
             user_tunnel_endpoint: None,
             status: Some(InterfaceStatus::Activated.to_string()),