+++ smartcontract/cli/src/link/dzx_create.rs
@@ -206,7 +206,8 @@ mod tests {
                 ip_net: "10.2.0.1/24".parse().unwrap(),
                 node_segment_idx: 0,
                 user_tunnel_endpoint: true,
-                ..Default::default()
+                // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+                // ..Default::default()
             }
             .to_interface()],
             max_users: 255,
@@ -240,7 +241,8 @@ mod tests {
                 ip_net: "10.2.0.2/24".parse().unwrap(),
                 node_segment_idx: 0,
                 user_tunnel_endpoint: true,
-                ..Default::default()
+                // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+                // ..Default::default()
             }
             .to_interface()],
             max_users: 255,
@@ -274,7 +276,8 @@ mod tests {
                 ip_net: "10.2.0.3/24".parse().unwrap(),
                 node_segment_idx: 0,
                 user_tunnel_endpoint: true,
-                ..Default::default()
+                // TODO: Uncomment this in next phase of InterfaceV2 rollout.
+                // ..Default::default()
             }
             .to_interface()],
             max_users: 255,